SBERDOMA-940 - brought the analytics view into separate components & add abstract data mapper class
