+++ apps/condo/domains/common/constants/style.js
@@ -29,6 +29,12 @@ const lightRed = '#FFF1F0'
 const brightRed = '#F5222D'
 const whiteTranslucent = 'rgba(255, 255, 255, 0.9)'
 
+const COLOR_SET = [blue[5], green[5], red[4], gold[5], volcano[5], purple[5],
+    lime[7], sberGrey[7], magenta[5], blue[4], gold[6], cyan[6],
+    blue[7], volcano[6], green[5], geekblue[7], sberGrey[7], gold[7],
+    magenta[7], yellow[5], lime[7], blue[8], cyan[5], yellow[6],
+    purple[7], lime[8], red[6]]
+
 const colors = {
     sberDefault: generateCustomColorPalette(sberBlue, sberSecondaryBlue),
     sberPrimary: generateCustomColorPalette(sberGreen, sberSecondaryGreen),
@@ -112,4 +118,5 @@ module.exports = {
     antGlobalVariables,
     shadows,
     transitions,
+    COLOR_SET,
 }
