+++ apps/condo/domains/ticket/constants/common.js
@@ -1,7 +1,9 @@
 const MAX_TICKET_REPORT_COUNT = 1000
 const PDF_REPORT_WIDTH = 1600
+const DATE_DISPLAY_FORMAT = 'DD.MM.YYYY'
 
 module.exports = {
     MAX_TICKET_REPORT_COUNT,
     PDF_REPORT_WIDTH,
+    DATE_DISPLAY_FORMAT,
 }
