+++ apps/condo/domains/ticket/utils/helpers.ts
@@ -5,7 +5,9 @@ import { LOCALES } from '@condo/domains/common/constants/locale'
 import moment, { Moment } from 'moment'
 import { ParsedUrlQuery } from 'querystring'
 import { Ticket, TicketAnalyticsGroupBy, TicketStatus, TicketStatusWhereInput, TicketWhereInput } from '../../../schema'
-import { ticketSelectTypes, viewModeTypes } from '../components/TicketChart'
+import TicketChart, { TicketSelectTypes, ViewModeTypes } from '../components/TicketChart'
+import { TableColumnsType } from 'antd'
+import { DATE_DISPLAY_FORMAT } from '@condo/domains/ticket/constants/common'
 
 export const getTicketCreateMessage = (intl, ticket) => {
     if (!ticket) {
@@ -375,8 +377,8 @@ export type ticketAnalyticsPageFilters = {
 interface IFilterToQuery {
     (
         filter: ticketAnalyticsPageFilters,
-        viewMode: viewModeTypes,
-        ticketType: ticketSelectTypes
+        viewMode: ViewModeTypes,
+        ticketType: TicketSelectTypes
     ): { AND: TicketWhereInput['AND'], groupBy: TicketAnalyticsGroupBy[] }
 }
 
@@ -408,3 +410,115 @@ export const filterToQuery: IFilterToQuery = (filter, viewMode, ticketType) => {
 
     return { AND, groupBy }
 }
+
+export const ticketChartDataMapper = new TicketChart({
+    line: {
+        chart: (viewMode, data) => {
+            const axisLabels = Array.from(new Set(Object.values(data).flatMap(e => Object.keys(e))))
+            const legend = Object.keys(data)
+            const series = []
+            Object.entries(data).map(([groupBy, dataObj]) => {
+                series.push({
+                    name: groupBy,
+                    type: viewMode,
+                    symbol: 'none',
+                    stack: groupBy,
+                    data: Object.values(dataObj),
+                    emphasis: {
+                        focus: 'none',
+                        blurScope: 'none',
+                    },
+                })
+            })
+            const axisData = { yAxis: { type: 'value', data: null }, xAxis: { type: 'category', data: axisLabels } }
+            const tooltip = { trigger: 'axis', axisPointer: { type: 'line' } }
+            return { series, legend, axisData, tooltip }
+        },
+        table: (viewMode, data, restOptions) => {
+            const dataSource = []
+            const { translations, filters } = restOptions
+            const tableColumns: TableColumnsType = [
+                { title: translations['address'], dataIndex: 'address', key: 'address', sorter: (a, b) => a['address'] - b['address'] },
+                {
+                    title: translations['date'],
+                    dataIndex: 'date',
+                    key: 'date',
+                    defaultSortOrder: 'descend',
+                    sorter: (a, b) => moment(a['date'], DATE_DISPLAY_FORMAT).unix() - moment(b['date'], DATE_DISPLAY_FORMAT).unix(),
+                },
+                ...Object.entries(data).map(([key]) => ({ title: key, dataIndex: key, key, sorter: (a, b) =>a[key] - b[key] })),
+            ]
+            const uniqueDates = Array.from(new Set(Object.values(data).flatMap(e => Object.keys(e))))
+            uniqueDates.forEach((date, key) => {
+                const restTableColumns = {}
+                Object.keys(data).forEach(ticketType => (restTableColumns[ticketType] = data[ticketType][date]))
+                let address = translations['allAddresses']
+                const addressList = get(filters, 'addresses')
+                if (addressList && addressList.length) {
+                    address = addressList.join(', ')
+                }
+                dataSource.push({ key, address, date, ...restTableColumns })
+            })
+            return { dataSource, tableColumns }
+        },
+    },
+    bar: {
+        chart: (viewMode, data) => {
+            const series = []
+            const axisLabels = Array.from(new Set(Object.values(data).flatMap(e => Object.keys(e))))
+            const legend = Object.keys(data)
+            Object.entries(data).map(([groupBy, dataObj]) => {
+                series.push({
+                    name: groupBy,
+                    type: viewMode,
+                    symbol: 'none',
+                    stack: 'total',
+                    data: Object.values(dataObj),
+                    emphasis: {
+                        focus: 'self',
+                        blurScope: 'self',
+                    },
+                })
+            })
+            const axisData = { yAxis: { type: 'category', data: axisLabels }, xAxis: { type: 'value', data: null } }
+            const tooltip = { trigger: 'item', axisPointer: { type: 'line' } }
+            return { series, legend, axisData, tooltip }
+        },
+        table: (viewMode, data, restOptions) => {
+            const { translations, filters } = restOptions
+            const dataSource = []
+            const tableColumns: TableColumnsType = [
+                { title: translations['address'], dataIndex: 'address', key: 'address', sorter: (a, b) => a['address'] - b['address'] },
+                ...Object.entries(data).map(([key]) => ({ title: key, dataIndex: key, key, sorter: (a, b) => a[key] - b[key] })),
+            ]
+            const restTableColumns = {}
+            const addressList = get(filters, 'addresses')
+            const aggregateSummary = addressList !== undefined && addressList.length === 0
+            if (aggregateSummary) {
+                Object.entries(data).forEach((rowEntry) => {
+                    const [ticketType, dataObj] = rowEntry
+                    const counts = Object.values(dataObj) as number[]
+                    restTableColumns[ticketType] = counts.reduce((a, b) => a + b, 0)
+                })
+                dataSource.push({
+                    key: 0,
+                    address: translations['allAddresses'],
+                    ...restTableColumns,
+                })
+            } else {
+                addressList.forEach((address, key) => {
+                    const tableRow = { key, address }
+                    Object.entries(data).forEach(rowEntry => {
+                        const [ticketType, dataObj] = rowEntry
+                        const counts = Object.entries(dataObj)
+                            .filter(obj => obj[0] === address).map(e => e[1]) as number[]
+                        tableRow[ticketType] = counts.reduce((a, b) => a + b, 0)
+                    })
+                    dataSource.push(tableRow)
+                })
+            }
+            return { dataSource, tableColumns }
+        },
+
+    },
+})
